+++ flutter_inappwebview_web/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 1.1.0
  
+- Migrates package to `package:web`.
 - Updates minimum supported SDK version to Flutter 3.24/Dart 3.5.
 
 ## 1.0.8